[frontend] add filter into timeline component
